Second Commit: Added Cards and more
I added a battle scene where I added some sample cards and spells that can be dragged and dropped to attack, defend and also to unlock combo spells that do much more. the spells can be dragged on to the player area to function. That part will be added in the next update. For now the cards can be dragged onto each other to unlock combo spells. More coming soon.